+++ examples/401-PostInfoList.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Example how to get a list of street names.
+ * Example how to get a list of post info values.
  */
 
 use DigipolisGent\Flanders\BasicRegisters\BasicRegister;
@@ -13,7 +13,7 @@ use DigipolisGent\Flanders\BasicRegisters\Pager\Pager;
 
 require_once __DIR__ . '/bootstrap.php';
 
-printTitle('Get a list of the first 25 street names in Gent from the service.');
+printTitle('Get a list of the post info of (sub)municipalities from the service.');
 
 printStep('Create the API client configuration.');
 $configuration = new Configuration($apiEndpoint, $apiUserKey);
@@ -27,13 +27,20 @@ $client = new Client($guzzleClient, $configuration);
 printStep('Create the Service wrapper.');
 $service = new BasicRegister($client);
 
-printStep('List of street names:');
-$filters = new Filters(new MunicipalityNameFilter('gent'));
-$streetNames = $service->streetName()->list($filters, new Pager(0, 25));
+printStep('List of (sub)municipalities of %s:', $examplePostInfoName);
+$filters = new Filters(new MunicipalityNameFilter($examplePostInfoName));
+$postInfos = $service->postInfo()->list($filters, new Pager(0, 25));
 
-foreach ($streetNames as $streetName) {
-    /** @var \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetName $municipalityName */
-    printBullet('%s : %s', $streetName->streetNameId(), $streetName);
+foreach ($postInfos as $postInfo) {
+    /** @var \DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoInterface $postInfo */
+    printBullet('%s', $postInfo);
+
+    if ($postInfo->postInfoNames()->hasSubMunicipalities()) {
+        foreach ($postInfo->postInfoNames() as $geographicalName) {
+            /** @var \DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName $geographicalName */
+            printText('    - %s', $geographicalName);
+        }
+    }
 }
 
 printFooter();