+++ src/Normalizer/FromJson/Post/PostInfoNormalizer.php
@@ -4,11 +4,10 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Post;
 
-use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Geographical\GeographicalNamesNormalizer;
 use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\IdExtractor;
-use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfo;
 use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoInterface;
 
 /**
  * Normalizes JSON data into a PostInfo value.
@@ -20,33 +19,16 @@ final class PostInfoNormalizer
      *
      * @param object $jsonData
      *
-     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfo
+     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoInterface
      */
-    public function normalize(object $jsonData): PostInfo
+    public function normalize(object $jsonData): PostInfoInterface
     {
         $idExtractor = new IdExtractor();
+        $postInfoNamesNormalizer = new PostInfoNamesNormalizer();
 
         return new PostInfo(
             new PostInfoId($idExtractor->extractObjectId($jsonData)),
-            $this->extractPostGeographicalNames($jsonData)
+            $postInfoNamesNormalizer->normalize($jsonData->postnamen)
         );
-    }
-
-    /**
-     * Extract the post geographical names from the json data.
-     *
-     * @param object $jsonData
-     *
-     * @return GeographicalNames
-     */
-    private function extractPostGeographicalNames(object $jsonData): GeographicalNames
-    {
-        $postNames = [];
-        foreach ($jsonData->postnamen as $wrapper) {
-            $postNames[] = $wrapper->geografischeNaam;
-        }
-
-        $geoGraphicalNamesNormalizer = new GeographicalNamesNormalizer();
-        return $geoGraphicalNamesNormalizer->normalize($postNames);
     }
 }