+++ tests/Normalize/FromJson/Post/PostInfosNormalizerTest.php
@@ -6,10 +6,10 @@ namespace DigipolisGent\Tests\Flanders\BasicRegisters\Normalize\FromJson\Municip
 
 use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\Post\PostInfosNormalizer;
 use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
-use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfo;
 use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfos;
 use PHPUnit\Framework\TestCase;
 
@@ -76,7 +76,7 @@ EOT;
         $expected = new PostInfos(
             new PostInfo(
                 new PostInfoId(9000),
-                new GeographicalNames(
+                new PostInfoNames(
                     new GeographicalName(
                         new LanguageCode('NL'),
                         'GENT'
@@ -85,7 +85,7 @@ EOT;
             ),
             new PostInfo(
                 new PostInfoId(9030),
-                new GeographicalNames(
+                new PostInfoNames(
                     new GeographicalName(
                         new LanguageCode('NL'),
                         'Mariakerke'