+++ tests/Value/Post/PostInfosTest.php
@@ -5,10 +5,11 @@ declare(strict_types=1);
 namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value\Post;
 
 use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
-use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfo;
 use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoInterface;
+use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfos;
 use PHPUnit\Framework\TestCase;
 
@@ -41,13 +42,13 @@ class PostInfosTest extends TestCase
      * @param int $identifier
      * @param string $name
      *
-     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfo
+     * @return \DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoInterface
      */
-    public function createPostInfo(int $identifier, string $name): PostInfo
+    public function createPostInfo(int $identifier, string $name): PostInfoInterface
     {
         return new PostInfo(
             new PostInfoId($identifier),
-            new GeographicalNames(
+            new PostInfoNames(
                 new GeographicalName(
                     new LanguageCode('NL'),
                     $name